update for supressing console flashes on windows
This modification was provided by Artem to me from someone else who was having same issues I was with the license helper subprocess causing console popups on licensing checks which was super annoying. This worked for me on windows and doesn't cause issues on Linux when tested. This applies the windows specific flags to hide these windows as outlined in https://docs.python.org/3.6/library/subprocess.html#windows-popen-helpers