update readme so code examples to be clear (#24)
Updated examples to match https://help.cryptolens.io/examples/key-verification so the there is not mixed cases where res and result is used and token examples show the == at the end like is needed and is reflected on help examples. This caused me great confusion till i trial and error tested and found that the == in the auth token was needed.  Snippets are now also easier to use to reuse and build code for noobs vs mismatched RSAPubKey and pubKey used.